requirements-lock.txt
# Pinned dependencies from working environment (2026-02-06)
# Use: pip install -r requirements-lock.txt
# To update: pip freeze > requirements-lock.txt
#
# Core MCP dependencies
mcp==1.25.0
pyyaml==6.0.3
python-dotenv==1.2.1
watchdog==6.0.0

# Discord bot
discord.py==2.6.4

# LLM providers
google-generativeai==0.8.6
google-genai==1.62.0
anthropic==0.78.0
httpx==0.28.1
ollama==0.6.1

# Transitive dependencies (pinned for reproducibility)
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
multidict==6.7.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.3.1
python-multipart==0.0.21
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
tenacity==9.1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
